// src/types/Champion.ts
// 챔피언 데이터를 위한 타입 정의

export interface Champion {
  id: string;
  key: string;
  name: string;
  title: string;
}
export interface ChampionList {
  [key: string]: ChampionData;
}

export type ChampionBase = Pick<Champion, "id" | "name" | "title" | "key">;
export interface ChampionData extends Champion {
  version: string;

  blurb: string;
  info: {
    attack: number;
    defense: number;
    magic: number;
    difficulty: number;
  };
  image: {
    full: string;
    sprite: string;
    group: string;
    x: number;
    y: number;
    w: number;
    h: number;
  };
  tags: string[];
  partype: string;
  stats: {
    hp: number;
    hpperlevel: number;
    mp: number;
    mpperlevel: number;
    movespeed: number;
    armor: number;
    armorperlevel: number;
    spellblock: number;
    spellblockperlevel: number;
    attackrange: number;
    hpregen: number;
    hpregenperlevel: number;
    mpregen: number;
    mpregenperlevel: number;
    crit: number;
    critperlevel: number;
    attackdamage: number;
    attackdamageperlevel: number;
    attackspeedperlevel: number;
    attackspeed: number;
  };
}